move to processor subdirectory
